make ReactiveConnectionPool a welldefined extension point
- introduce the 'pool' package
- customizable via the Hibernate Services stuff
