+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/boot/impl/ReactiveServiceRegistryInitializer.java
@@ -2,7 +2,7 @@ package org.hibernate.reactive.boot.impl;
 
 import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
 import org.hibernate.reactive.jpa.impl.ReactivePersisterClassResolverInitiator;
-import org.hibernate.reactive.service.initiator.ReactiveConnectionPoolInitiator;
+import org.hibernate.reactive.pool.impl.SqlClientPoolInitiator;
 import org.hibernate.reactive.service.initiator.DummyConnectionProviderInitiator;
 import org.hibernate.reactive.service.initiator.ReactiveJdbcEnvironmentInitiator;
 import org.hibernate.reactive.service.initiator.ReactiveQueryTranslatorFactoryInitiator;
@@ -20,9 +20,9 @@ public class ReactiveServiceRegistryInitializer implements ServiceContributor {
 	public void contribute(StandardServiceRegistryBuilder serviceRegistryBuilder) {
 		serviceRegistryBuilder.addInitiator( DummyConnectionProviderInitiator.INSTANCE );
 		serviceRegistryBuilder.addInitiator( ReactiveJdbcEnvironmentInitiator.INSTANCE );
-		serviceRegistryBuilder.addInitiator( ReactiveConnectionPoolInitiator.INSTANCE );
 		serviceRegistryBuilder.addInitiator( ReactiveTransactionCoordinatorBuilderInitiator.INSTANCE );
 		serviceRegistryBuilder.addInitiator( ReactivePersisterClassResolverInitiator.INSTANCE );
 		serviceRegistryBuilder.addInitiator( ReactiveQueryTranslatorFactoryInitiator.INSTANCE );
+		serviceRegistryBuilder.addInitiator( SqlClientPoolInitiator.INSTANCE );
 	}
 }
