+++ hibernate-reactive-core/src/main/java/org/hibernate/reactive/pool/ReactiveConnectionPool.java
@@ -1,4 +1,4 @@
-package org.hibernate.reactive.service;
+package org.hibernate.reactive.pool;
 
 import org.hibernate.service.Service;
 
@@ -8,7 +8,9 @@ import java.util.concurrent.CompletionStage;
  * A Hibernate {@link Service} that provides access to pooled
  * {@link ReactiveConnection reactive connections}.
  *
- * @see SqlClientPool
+ * A program may integrate a custom {@link ReactiveConnectionPool}
+ * with Hibernate Reactive by contributing a new service using a
+ * {@link org.hibernate.boot.registry.StandardServiceInitiator}.
  */
 public interface ReactiveConnectionPool extends Service {
