+++ hibernate-reactive-core/src/test/java/org/hibernate/reactive/BaseMutinyTest.java
@@ -11,8 +11,8 @@ import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.reactive.containers.PostgreSQLDatabase;
 import org.hibernate.reactive.mutiny.Mutiny;
-import org.hibernate.reactive.service.ReactiveConnection;
-import org.hibernate.reactive.service.ReactiveConnectionPool;
+import org.hibernate.reactive.pool.ReactiveConnection;
+import org.hibernate.reactive.pool.ReactiveConnectionPool;
 import org.hibernate.reactive.stage.Stage;
 import org.junit.After;
 import org.junit.Before;
