+++ hibernate-reactive-core/src/test/java/org/hibernate/reactive/BaseReactiveTest.java
@@ -7,8 +7,8 @@ import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.reactive.containers.PostgreSQLDatabase;
-import org.hibernate.reactive.service.ReactiveConnection;
-import org.hibernate.reactive.service.ReactiveConnectionPool;
+import org.hibernate.reactive.pool.ReactiveConnection;
+import org.hibernate.reactive.pool.ReactiveConnectionPool;
 import org.hibernate.reactive.stage.Stage;
 import org.junit.After;
 import org.junit.Before;
